Add a disassembler for methods
This will be used to create dynamic methods from the provided methods and functions at the IL level.
It parses the bytes of the method body into an array of opcodes and operands for the 'Call' or 'Callvirt' opcodes. These operands (metadata tokens) are then used to look up the correct member reference in the metadata table. The MethodInfo objects are then resolved using this metadata and mapped to the instruction position of the corresponding call opcode.
These MethodInfo objects are then used when emitting a dynamic method.

TODO: Handle TypeDefs, TypeSpecs, MethodsDefs and (Fields?)